i2c: tegra: fix usage of PM runtime framework
runtime_idle is a callback to check to see if a device should suspend,
it is not the function that should do any suspend activity.
runtime_suspend is the correct callback.

pm_runtime_put is preferred to pm_runtime_put_sync, since we do not
need the subsequent clk_disable to return synchronously.

Bug 886799

Change-Id: I9c5216dedd746773a26147ba1feedcff43aa4834
Signed-off-by: Tom Cherry <tcherry@nvidia.com>
Reviewed-on: http://git-master/r/71081
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Dan Willemsen <dwillemsen@nvidia.com>